Fix a regression in dcmf-conduit (BG/P)
The GASNETE_BARRIERFLAG_UNNAMED support was leading to deadlock on only
the OPT variant of several test on all non-pthreaded configurations on
BlueGene/P.  Specifically (at least) the following tests:
      bugzilla/locktest
      gwu-upc-io/perf_local
      intrepid/test17
      intrepid/test18
      intrepid/test28
      upcr-tests/locks
There are other suspicious OPT-specific tests that share the same root
cause, but I've not listed them here.


The root of the problem is that the code being used to leverage to h/w
barrier suport does NOT poll the AM queues.  This code is disabled for
PAR builds because it is not thread-safe.  It is unused in a DBG build
because we used named barriers for error checking.

Since we don't poll for AMs when in a runtime-internal barrier, even a
very simple test like upcr-tests/locks can deadlock when, for instance,
a thread has entered the final (post-main) barrier and other threads must
complete one or more AM-based transations (such as upc_unlock) with that
thread before they can reach the same barrier.

Ideally it might be possible to fix the code to ensure that we can poll
both the h/w barrier and AMs.  HOWEVER, both IBM and Argonne consider the
BlueGene/P system to be end-of-life.  So, I've taken the simple approach
of disabling the new code that lead to this regression.