Apply a patch supplied by Mellanox
This commit fixes mxm2 support by accomodating an API change made in
their final version (released just recently).
An unused struct member is also removed as related clean-up

NOTE: This patch is currently under consideration for inclusion in
the pending release.  This was expected to be the first release to
include mxm2 support, but without this patch that won't be the case.
Tests are queued on Mellanox's DMZ cluster...