Merge to HEAD of BG_COMMONHEAP_branch
This commit merges the branch containing the new "GHEAP" implementation
of PSHM, and the supporting scripts/docs to enable it by default on BG/Q.

The Global HEAP implementation of PSHM does not contain any code which is
specific to BG/Q.  The implementation is suitable for any runtime in which
the virtual addresses can be freely passed among (and validly used by) all
of the processes on a given compute node.  This could potentially be the
case on any system with real mode addressing which doesn't have protection
between address spaces (or one can disable the protection).  It is also
worth noting that mmap() support is not required.

The necessary addressing property is true on BG/Q *only* if one sets
BG_MAPCOMMONHEAP=1 in the compute node environment.  So, the spawner
scripts have a little bit of new logic to set it automatically.

The code currently uses posix_memalign()/free() to alloc/free pages for
segments.  However, that logic is easily modified to use other mechanisms.
In particular, one could probably implement something that used SHMEM for
the heap allocation on a system with shmem_ptr() support for a hybrid with
an additional network conduit (e.g. Altix nodes connected w/ IB).
Additioanlly, introducing computable offsets would require only small
localized changes to the existing code.