Embed version info in tarballs generated by "make dist"
We now embed the "git describe" output in TWO places in a distribution tarball:
+ Once "in plain sight" in version.git
+ Again "hidden" in the generated configure script (but NOT in configure.in,
  so that file will continue to exactly match the in the repo)

This commit also updates the release procedure to ensure the final tarballs are
generated AFTER the release tag is applied to get the final versions.

"For good measure" the info hidden in the configure script now distinguishes
the full and tools-only configure scripts.