Remove atomics for Cray T3E and NEC SX series
This commit removes the explicit "use generics" checks for the
Cray T3E and NEC SX series platforms, neither of which we have
been able to test for a signifcant period of time.

This also removes TWO unused implementations for SX series.

The intent is that the code will soon be restructured such that
use of generics will be the default.  Thus, anybody who still
uses either platform is not losing anything.