Ensure automatic fall-back to generic atomics
Prior to this commit an unknown CPU would result in an #error.
Now we default to using the generics by default.
This is "friendlier" to those porting to a new system, and since
our tests complain loudly about the use of the generic atomics,
there is no risk that a porter will forget the need (desire) for
native atomics.

I am *not* making the corresponding change for the case of an
unknown (or unsupported) compiler for a known architecture.  That
can still be debated.