Remove remaining HP/UX support
This commit removes both OS-specific and compiler-specific support for
HP/UX and the HP compiler family.

There is no evidence I can see to suggest that we ever supported HP's
compiler for x86-64 (which might actually have been Compaq's compiler,
which we also have never claimed to support on x86-64).