Distinguish sub v. subsub section in gasnet.texi
This commit splits the FNH and FN macros into FNHS, FNHSS, FNS and FNSS.
Prior to this commit all functions were specified at the subsubsection
level, but now the *S macros define some at the subsection level, while
the *SS macros define the remainder at the subsubsection level.

This change is needed to keep the makeinfo-5 parser happy with the
definitions (still incomplete) of @menus.