Avoid redefinition of @comma
This commit prevents us from redefining the @comma macro when using
makeinfo-5, while retaining its explicit definition with makeinfo-4.

The obvious mechanism (@ifcommandnotdefined comma) only began with
the release of makeinfo-5.0, but the texinfo manual documents a way
to test for the existance of @ifcommandnotdefined itself, which is
sufficient for our needs.