Eliminate unnecessary RMB in ibv dynamic connect
As described in the new comment, there is actually no need for
an RMB at exit from the "while (NULL == result)" loop, even on
weakly-ordered PPC.  This was actually only ever needed on the
DEC Alpha (which GASNet no longer supports).

See also:
Documentation for "read_barrier_depends()" in the Linux kernel.