Fix bug 3200
This commit resolves
   Bug 3200 - startup failures on Cray systems with craype-2.x

The fix is to use the -whole-archive linker flag for libhugetlbfs
as documented in Cray's intro_hugepages man page (but contrary to
the instructions in their aprun man page).

The root cause of the problem is that in craype-1.x the 'cc'
compiler wrapper was automatically transforming a lone instance of
'-lhugetlbfs' to '-Wl,--whole-archive,-lhugetlbfs,--no-whole-archive'
but in craype-2.x no longer performs that transformation (my thanks
to Zhengji Zhao of NERSC for that diagnosis).