Cray cc and CC have deprecated -target=linux
This commit removes "-target=linux" from the definitions of CC and CXX
in the cross-configure script for the Cray XE.

Cray's current compiler wrappers were translating this value to
"-target=compute_node", which is the default and also the only fully
supported target.  The wrappers on our XC already complain about
"-target=linux" and there is a reasonable chance the XE will
eventually, too.  So, this change is preemptive.