Yet more spaces after string literals for C++11
The sites for this round of changes was found using:

for x in `find . -name \*.h`; do
  echo @$x;
  perl -n -e 'print if (/^([^"]*"[^"]*[^\\]")+[_A-Za-z]/);' -- $x;
done

Which looks for the first character of an identifier (letter or
underscore) immediately following a double quote which is the
2nd, 4th, etc. on a given line.  It ignores properly nest pairs
of backslash-prefixed double-quotes which fall between bare
double-quotes. So "this is a \"properly nested\" example" does
NOT match at the 'p' in 'properly'.

In order to ensure the same perl regex will be useful in the
future, I also changed the two following false positives:

@./gasnet_atomic_bits.h
          : "=&b"(result), "=m" (v->ctr)        /* constraint b = "b"ase register (not r0) */
  This is a comment and thus NOT an actual problem.
  I changed double-quotes to single-quotes to avoid future matches.

@./other/fca/gasnet_fca.h
    fprintf(stderr,"%s:%d - %s(): rank %d: "format"\n", __FILE__, __LINE__, __FUNCTION__, \
  This is not an issue because this a private header.
  However, I added spaces around 'format' anyway.