Introduce _GASNETI_PARSEQ for internal use
This commit introduces preprocessor constant _GASNETI_PARSEQ which has
a value of either _PAR or _SEQ, depending on the value of
GASNETI_THREADS.  The intended use is as a suffix, probably added
using _CONCAT(), for identifiers used with types, functions and macros
which have both parallel and sequential variants.  For instance the
"weak" atomics use either a parallel or sequential version depending
on GASNETI_THREADS.  Proper use of _GASNETI_PARSEQ will aide in
producing cleaner code for variants like the weak atomics.