_PAR and _SEQ variants of gasneti_semaphore_t
This commit renames the previous implementation of gasneti_semaphore_t
to have a _PAR suffix, and adds a new (sequential) implementation with
the same types and functions, but with a _SEQ suffix.  Then the
original identifiers are mapped to either the _SEQ or _PAR versions as
determined by the value of _GASNETI_PARSEQ.

This is the first of hopefully several cleanups that will be
implemented based on this idea, first proposed by Eric Hoffman for use
in reducing the complexity of the atomics.