_PAR and _SEQ variants of gasneti_lifo_head_t
This commit moves ibv's conduit-specific implementation to
gasnet_syncops.h with a _SEQ suffix and renames the prior
implementation with a _PAR suffix.  The original, un-suffixed, names
are mapped to either the _PAR or _SEQ variant based on
_GASNETI_PARSEQ.

Additonally, the ibv-specific gasnetc_lifo_* code is mapped to the
_PAR or _SEQ variants in gasnet_syncops.h using the _CONCAT mechanism
to reduce needless duplication.