Reorg the name-shifting of PAR-v-SEQ syncops
This commit changes how the new logic from the previous few commits
is implemented to allow the general and conduit-specific paths to
use the same CONCAT-based macros.

The result is longer lines, but fewer of them.  However, the somewhat
"tricky" bit regarding order of concatenation (to avoid symbol
conflict on the intermediate value) needs only to be dealt with ONCE
by the generic code, and not be each conduit (or subsystem) which may
want to manually select between the SEQ and PAR variants.