Add logic to build ops via gasneti_atomic_fetchadd
Prior to the commit we had logic could build several ops (inc, dec,
dec-and-test, add and sutract) automatically from platform-specific
atomics including UNfenced fetch-and-add, but not from the fully
fenced fetch-and-add.  This commit adds that missing logic to the
conditional-compliation ladder and additionally reorders the
logic to define the fixed-width operations to give precedence to
the fully-fenced variant of fetch-and-add if (unlikely) both are
defined.