Reimplement fence-handling for atomics
This commit revises the handling of fences in atomics to hopefully
be far more readable and maintainable.  One level of "magic" is
replaced by having each implementation override any fencing macros
which should not use the default values.  This replace the previous
mechanism in which an implementation defined one or more bit masks
that were used at pre-process time to construct the appropriate
non-default values for the macros that are now set directly.  The
resulting code is significantly shorter.