Restructure weak atomics implementation
This commit restructures the way the "weak" atomics are implemented so
that the three "nonatomic" implementations ("nonatomic",
"nonatomic32", and "nonatomic64") are always compiled.  Now the weak
atomics are just an alias to either the "non"-prefixed or unprefixed
versions.  This aliasing is done using the concat-based approach which
was originally proposed by Eric Hoffman as a way tohelp clean-up this
code.

One purpose behind this work is to expose the "nonatomic"
implementations to ibv-conduit (and maybe others) which currently
implement their own versions (for a case where GASNETI_THREADS is
defined but the need for certain operations to be atomic is dependent
on whether a SPECIFIC thread is enabled).