Fix two issues in generic atomics
This commit fixes two issues that crept into the generic (muted-based)
atomic during the recent revision of the fencing logic.

1) The mask to avoid duplicating the fences implied by the lock and
unlock were incorrect.  This rendered the de-duplication ineffective
but could not have resulted in missing fences (thus a peformance bug
not a correctness one).  This commit applies the correct masks.

2) The simplification of the alignment check removed the assignment to
a dummy variable, leading to decl-after-stmt warnings from gcc from
ONLY the generic atomics because of their declaration of a lock
pointer variable.  This commit moves the declaration of the lock
pointer into a newly created "prologue" piece of the fenced atomic
templates.  The prologue is currently empty for all but generic
atomics, but has been placed after the last declaration in the
templates so that it may safely contain both declarations and
statements.