Fix for bug introduced in commit f1dfbdd
This commit resolves a bug introduced in commit f1dfbdd when trying to
resolve decl-after-stmt warnings in the generic atomics.  The problem
is that the generic atomic read code will try to use the lock type
(hsl or pthread) prior to its definition.  This is a fatal compilation
error, as opposed to the warning(s) it replaced.

The introduction of a single prologue was insufficient.  Instead
distinct set/read/rmw prologues are now defined, with the read case
defaulting to empty.