Change mechanism for non-default atomic fences
This commit moves the default definintions of the atomic fencing
macros to come before the platform-specific code, included from
gasnet_atomic_bits.h.  With this change the default definitions are
now unconditional, as opposed to protected with #ifndef.  As a result
the cases in gasnet_atomic_bits.h which need to provide custom
definitions must #undef the defaults before redefining them.  This
change results in a net savings of 9 lines currently, but when the
32-bit, 64-bit and custom atomics each have distinct fencing macros,
the difference will be more significant.