Simplify constuction of arithmetic atomic ops
This commit greatly simplifies how the arithmetic atomic ops (inc,
dec, dec-and-test, add and subtract) are constructed from either
fetch-and-add or add-and-fetch.  In particular the logic now defines
the arithmetics in terms of the fully-fenced add-and-fetch, which it
will constuct from the other supported options if needed.

Notably, the construction of add-and-fetch from compare-and-swap has
been removed.  No current platform needs it, and one could supply the
logic if needed.