Start update to 64-bit atomics logic
This commit begins the process of updating the 64-bit atomics logic in
gasnet_atomicops.h to match that for the 32-bit atomics.  It only does
the initial steps of changing a few existing definitions.  I have not
yet added any of the new logic required to support 64-bits as the
native atomic type.