Split fencing macros into three
This commit "triplicates" the fencing macros to have distinct groups
for the 32-bit, 64-bit and custom (if any) atomic types.

Additionally, the mechainsm for overridding the default definitions
has reverted to PRE-definition (vs RE-definition) and the grouping of
related overrides (must pre-define all or none from a group) has be
restored.

This change eliminates the need for "sneaky" direct implementation of
the fully-fenced 64-bit read and set operations on (32-bit) x86, and
generally makes the implementation of new atomic implementations less
tricky.