Fix multiple errors in "hybrid" 64-bit atomics
This commit corrects errors in the 64-bit swap and addfetch operations
as implemented for ILP32 on PPC64 hardware on Darwin.  This
configuration is known as "hybrid" atomics because the ABI doesn't
ensure 8-byte alignment of 64-bit type and we switch, at runtime,
between native and mutex-based atomics based on actual alignment.

The errors include a link failure, which clearly shows that I somehow
missed testing the final code in this configuration.  My bad!

Errors fixed:
+ 64-bit swap and fetchadd were returning int rather than uint64_t
+ the inline asm provides fetchadd, but C code invoked addfetch
+ update comment to match fetchadd-vs-addfetch change