Fix missing fences in decrement-and-test
This commit resolves failures of benchmarks/barrierperf observed on
bluedrop_gcc_64_pthr (a ppc64 system) in the 08Feb2014 nightly tests.

The failure occured because the new default construction of
dec-and-test from the fully-fenced add-and-fetch operation did not
allow for the implementation of the conditional fences.  This issue
left the pthread-level upc_notify() without an important ACQ after the
last pthread reached the dec-and-test.  The issue didn't appear in
GASNet-level testing, nor in UPC-level testing on x86/x86-64 systems.

This commit leaves the default implementation of dec-in-test in terms
of the same fully-fenced add-and-fetch operation, but adds the
fencing.