Follow up with fully-fenced dec-and-test
The previous commit simplified the logic for construction of
dec-and-test, but a the expense of ONLY constructing it automatically
from UNfenced operations.

This commit "repairs" the two platforms which define fully-fenced
fetchadd operations (gcc and cray __sync atomics), by providing the
trivial implementation.  Note that automatic construction in this
manner is not safe in general, but does work for these cases in which
the fetchadd includes a FULL fence independent of the flags argument.