Don't ignore registration failure in NDEBUG build
This commit ensures that all builds (not only a DEBUG build as before)
will fail with an informative error message if we fail to register the
segment in a SEGMENT_FAST configuration.  Prior to this commit the
check was only done in an assert-like macro that compiles to nothing
unless debugging is enabled.

Thanks to Rafael Larrosa Jiménez for reporting the issue.