Take more care with CFLAGS for Open64 & PathScale
Both opencc and pathcc support -Wdeclaration-after-statement, but
lack the corresponding -Wno- form.  This is a problem for any client
code which wants to compile C99 code without warnings, because they
have no (simple) mechanism by which to override GASNet's addition
of -Wdeclaration-after-statement to the CFLAGS in a debug build.
Of particular concern at the moment is the Clang-based "upc2c"
source-to-source translator.

This commit disables the inclusion od -Wdeclaration-after-statement
in the CCDEBUGFLAGS and CXXDEBUGFLAGS for Open64 and PathScale.