Factor and correct upward half-barrier
This commit replaces 14 individual instances of "optional in-barrier
on same tree as collective" with just two instances.  However, the
more important change is that one of the instances includes memory
fences not present prior to this commit.  These are necesary for
correct operation (expsicially on PPC with its relatively weak memory
model) when PSHM is in use.  For instance, with a IN_ALLSYNC
broadcast it is possible that the datum to be sent was written to
the root node's address space by another process just prior to start
of the collective.  That is exactly what testcoll's ALL/ALL tests
do and they were infrequently failing broadcast and scatter validation
because of the missing fences.