Allow "natural" use of atomic add/subtract
This commit updates the tools docs to state explicitly the nature of
the "platform dependent" result of out-of-range atomic addition and
subtraction and removes the prohibition against negative operands.
We had already made two's-complement explicit elsewhere in our docs
for atomics.

Additionally, an assertion which had prevented use of negative
operands (but only under rare configurations) is removed.