Ensure Gemini/Aries conduits always use hugepages
This commit makes minor changes to ensure that the GASNet segment
will be backed by hugepages when GASNETI_USE_HUGETLBFS is defined,
independent of GASNETI_PSHM.

It has always been the intention that hugepages are to be used for
the GASNet segment with Gemini and Aries conduits.  However, it
seems we've not been doing so for the case where PSHM has been
disabled, even though we went through some effort in configure
to ensure we had the necessary support in this case.

Thanks to Mike Bauer of Stanford both for identifying the problem
and for the initial version of the solution.

(cherry picked from commit 005c6df7a2490d002d2e0ad34ee3f7cb4ba92ad8)