Reorder network probes (ibv after mxm and ptl4)
As observed by Gary Funck, the previous ordering would lead to use
of portals4-conduit as the default network on any system with the
reference implementation of ptl4 (which runs over ibv).  That is
not an appropriate default since the performance and stability
of portals4-over-ibv are both lower than bare ibv.

The identical argument applies to the current mxm-conduit, though
in the future that may not be the case as Mellanox introduces new
hardware and firmware which may better support mxm.