Implement GASNETE_USE_LONG_PUTS setting for AMRef
This commit implements a new parameter, GASNETE_USE_LONG_PUTS,
controlling the behavior of the AM-based reference Extended API.
This setting defaults to 1 (to yield behavior identical to before this
commit).  However, seting to 0 (in gasnet_extended_fwd.h) will disable
the use of Long and LongAsync Requests for implementation of puts (and
thus use a long stream of Medium Requests for large xfers).

The main purpose of this knob is for experimental work on resilience
for which we need Extended API paths that use only send/recv for
use during checkpoint coordination.  Under that scenario (with
ibv-conduit) the native Extended API implementation would be used
for the normal case, and only during a quiesce/drain operation would
the AMRef code be used.

The Design comments are also corrected by this commit to use "<="
rather than "<" in psuedo-code for the switch from Med to Long.