Fix a (never observed) corner-case bug in extended-ref
By using an int for the chunk size in two loops, the AM-based
reference implementation of the Extended API could have malfunctioned
for the case of a conduit with a MaxLong{Request,Reply} larger than
INT_MAX (2GB-1 on all platforms we currently support).

FWIW: We already have a sanity check against 4G-1, to ensure the
chunksz value will always fit in an AM handler argument.