Factor AMRef-related config sanity checks
This commit implements gasnete_check_config_amref() which contains
the configuration sanity checks for the parameters related to the
AM-based reference implementation of the Extended API, which replaces
the checks that had propogated to some (but not all) conduits when
they cloned gasnet_extended.c.

Additionally, the new gasnete_check_config_amref() takes some care
only to validate the parameters related to portions of the AMRef
implementation which are actually enabled to be built.