Fix EINVAL at munmap w/ hugepages but w/o PHSM
A recent change to use huge pages for the segment on the Cray XE, XK
and XC systems modified only the mapping path but failed to make the
corresponding change to the UNmapping path.  This was unnoticed in
testing because the unmapping is not frequently exercised for
small-memory tests as in the GASNet test suite.

Note that this bug could only occur with Aries and Gemini conduits,
and then only if PSHM was disabled at configure time.

Thanks to Elliot Ronaghan at Cray for reporting the problem and
for assitance testing the solutuion.