Update cross-configure-help stack direction probe
This commit updates the probe for the direction of stack growth to
match the one recently committed to UPCR's configure script.  This
one is known to be more robust in the face of compiler optimizations
than the previous version.

I've also added a note in the instructions to discourage enabling
optimizations when building the cross-configure-helper, which is
probably the best solution to the stack growth probe.