Reduce time spent probing max pinnable memory
Rather than trying to pin as much memory as possible in 4M chunks, we
now try pinning in large chunks.  We initially grow the limit linearly
by the large chunk size, and then try 1/2, 1/4, 1/8, etc.  of the
chunk size summing the sucessful ones to get the total.

In the best case we pin the max (default 2/3 of physical mem) on the
first attempt.  If not, the probe is limited by the PIN_MAXSZ (or the
HCA's equivalent: max_mr_size).  There is a linear number of trials
proportional to the pinnable memory divided by the PIN_MAXSZ (which is
typically MUCH larger than the old 4M chunk size), plus an additional
number of trials logarithmic in PIN_MAXSZ.