Add support for PMI2 (incomplete)
This commit is a rewrite of Alex Mikheev's pull request #14.
However, it is incomplete: lacking a barrier implementation for PMI2.

I have changed Alex's logic to allow a user to select a specific PMI
version with --with-pmi-version={1,2} and to stop with an error
mesasge if the user's choice is not avaiable.