Fix a minor error in testtools
A (relatively new) warning from the gcc built from Intrepid's sources
shows that the wrong "not" operation was being used in this test.
The result is that the test was weaker than intended, passing if ANY
bit was non-zero rather than requiring that ALL bits be non-zero.

The corresponding 64-bit test was correct.