Eliminate multi-registration under multidomain
With this commit we no longer register the segment once per domain
when using multi-domain.  Instead single registration always done at
attach time is used by all threads regardless of domain.

Cray has confirmed that memory-registration is global to all Cdm
handles (and derived Cq and Endpoint objects) which use the same
(ptag, cookie) in the call to GNI_CdmCreate, and for the dev_id
passed to GNI_CdmAttach (though current Crays are single-nic).

If we define "domain" to be the cdm_handle, then Cray has also
confirmed that serialization is required for access to objects within
the same "domain", but is NOT required when operating on objects in
different ones.

There are also a limited number of calls which require no serialization
(of which I don't yet have a complete list).  One of these is
GNI_CqTestEvent(), which was used in a prior commit to reduce lock
contention when polling for AM arrivals in PAR builds.

These important facts all appear in print in:

SE Choi, H Prichard, J Shimek, J Swaro, Z Tiffany, and B Turrubiates.
An Implementation of OFI libfabric in Support of Multithreaded PGAS
Solutions.  In Proceedings of the 9th International Conference on
Partitioned Global Address Space Programming Models, September 2015.

One can also examine https://github.com/ofi-cray/libfabric-cray for
the corresponding Open Source Software, written by current and former
Cray employees, which utilizes the global nature of the memory
registration and the thread-safetly of access to distinct domains.