amudp: C/R logic in AMUDP_SPMDStartup and amudprun
This commit adds logic for (BLCR-based) checkpoint/restart logic to
AMUDP.  Now AMUDP applications can be checkpointed and restarted.
Currently checkpoints must be collective and self-initiated, and
restarts must be collective as well.

This commit "breaks" (temporarily) GASNet's udp-conduit in the sense
that with these changes it is not possible to to link udp-conduit
applications with BLCR support enabled.  Of course, if the BLCR
support is not enabled, then udp-conduit still works just fine.
However, changes to udp-conduit are comming soon.