Fix for ppc timer asm with recent clang
Since about POWER4 the mftb group of instructions have been deprecated
in favor of mfspr with the reserved register numbers 268 and 269.  The
llvm integrated assembler for ppc is issuing deprecation warnings for
the gasnet timer code that uses the old instructions.

This commit (re)implements the tick reading code using the mfspr
instructions.  Currently this is dependent on use of clang, because
that is the only place the older instructions are a problem.  However,
one would ideally have a configure probe to determine which
instructions to use (or a compiler built-in when available).