Suppress a clang warning
Clang warns that abs(exp) is pointless when exp has type (unsigned
long long) as in the code modified by this commit.  I don't agree 100%
with that warning, since the conversion to (int) for parameter passing
can produce a signed value.  Even just making an explicit cast appears
sufficient to silence clang's warning.

To ensure the original intent of the code, this commit adds a cast to
a signed type (long long) before the subtraction.  To allow for large
differences, abs() has been changes to labs().  Note that C89 includes
labs(), but not llabs().