Long-overdue inlined extended API for smp-conduit
This commit adds a conduit-specific extended API implementation to
smp-conduit.

When fist conceived, the smp-conduit was a single process with the
option of using pthreads.  Since the gasnet_extended.h header has
logic to perform loopbacked Put and Get, they were always at least
possible to inline.  However, with the introduction of PSHM support
the Put and Get between processes was taken over by code inside the
(AM based) reference extended implementation and the ability to inline
was consequentially lost.  This commit restores the ability to fully
inline the extened API for smp-conduit.

In the future, this is likely to be removed in favor of a layer
*between* gasnet_{get,put,...etc...} and their gasnete_* equivalents.
That layer (in a header to ensure it is inlinable) will contain the
PSHM-bypass code, and will replace the use of GASNETI_CHECKPSHM_*
macros in each individual conduit.