Resolve bug #3260
On a system with multiple MIC's (aka Xeon-Phi's) our logic for
constructing the nodemap from the IB adapter's LID is flawed, since
every MIC will report the same LID (the host's).

The fix is to NOT use the IB LID for nodemap construction when on a
MIC system (PLATFORM_ARCH_MIC).  Strictly speaking this is over-kill
since it does not distinguish single-MIC systems or self-hosted MICs.
However, the penalty for this "over-kill" is just an additional
4-byte bootstrapExchange done once at job startup.