Remove unnecessary special case for XPMEM
This commit removes an #if-branch specific to PSHM-over-XPMEM which
tried to distinguish between the cases with and without hugetlbfs.
However, the cases for PSHM-over-POSIX and -FILE use a call to just
gasneti_munmap() with ALSO contains an #if-branch for the presence
or absence of hugetlbfs support.

So, XPMEM is now handled by the same case as POSIX and FILE.