Update ibv-conduit to support new XRC API
The XRC support used in GASNet's ibv-conduit so far has been an API
developed by Mellanox.  However, there is now an official OFED API for
XRC and it is different.

This commit adds logic for use of the OFED XRC API, in *addition* to
the Mellanox one.  There is also new configure logic to determine
which API is present in libverbs, and to use the OFED one if it is
available (including the case that both are present, as is the case in
current releases).