Implement gasneti_nodegrp_t for supernode "stats"
This commit implements a struct which gathers together the information
about the supernode which has so far been in 4 scalars and 1 array.
The same data can be applied to other groups, such as host for which
there is an immediate need (to correct the mmapLimit probe when the
supernode is smaller than the host due to GASNET_SUPERNODE_MAXSIZE).
This may later be applied to other architectural levels such as core,
socket, L3-cache, etc.

The following #define's preserve the legacy global variables:
  #define gasneti_nodemap_local        gasneti_mysupernode.nodes
  #define gasneti_nodemap_local_count  gasneti_mysupernode.node_count
  #define gasneti_nodemap_local_rank   gasneti_mysupernode.node_rank
  #define gasneti_nodemap_global_count gasneti_mysupernode.grp_count
  #define gasneti_nodemap_global_rank  gasneti_mysupernode.grp_rank