Correct gasnet_nodeinfo_t.supernode for non-PSHM
Prior to this commit the 'supernode' field of the nodeinfo was being
initialized the same regardless of whether PSHM was enabled or not.
This was potentially "misleading" for the client in a non-PSHM build
because the same 'supernode' normally indicates the ability to share
memory.  He was long had the 'node' field in the nodeinfo to convey
the same compute not independent of the ability to share memory.

This commit ensures that for a non-PSHM build we effectively have the
same effect as setting GASNET_SUPERNODE_MAXSIZE=1.