ssh-spawer: honor GASNET_NODEFILE env var
This commit adds GASNET_NODEFILE to the environment variables which
the ssh-based spawner will examine.  It is given precedence after the
variables GASNET_SSH_{NODEFILE,SERVERS} (since it is "less specific")
but before all of the variables specific to the various batch systems.

This change will allow for better integration with upcrun and other
scripts.  Prior to this change an ibv-conduit job would look at the
GASNET_NODEFILE environment variable only if the mpi-based spawner was
being used, and the GASNET_SSH_{NODEFILE,SERVERS} env vars only if the
ssh-based spawner was used.  Now, the environment variable
GASNET_NODEFILE is honored by both spawners.