atomics: safer x86 asm for 64-bit atomics w/ PIC
When compiling for PIC (position independent code), some x86 compilers
will disallow use of EBX in inline ASM, while others are smart enough
to save/restore it as necessary.  To deal with the cases where the
compiler considers EBX to be reserved, we have a distinct version of
the 64-bit atomics for x86 (because cmpxchg8b must use EBX).

Prior to this commit, our inline ASM for the EBX-is-reserved case was
written with the assumption that the compiler would never pick EBX for
an "r" constraint.  However, we have now encountered at least one case
where our heuristic for deciding when to use this case is incorrect.
So, this commit makes small changes to one of the operand constraints
to prevent alloction of EBX in a manner that would render the inline
ASM incorrect.  Specifically, we now list EDI and ESI instead of
allowing any register, since those are the only two left available.