atomics: more conservative use of PIC-safe x86 asm
When an x86 compiler reserves EBX (as the GOT pointer for PIC code) we
must use a specialized implementation of the 64-bit atomics that does
some extra work to save/restore EBX manually.  Prior to this commit,
we activated this special-case when either (1) a configure probe has
determined that the compiler will not allow "b" as an operand
contraint in inline asm, or (2) when __PIC__ is defined with a
compiler "known" to reserve EBX.

Somewhere in GCC-5 development, gcc began scheduling EBX "normally"
even when generating PIC code.  So, our list of compilers "known" to
reserve EBX is no longer accurate, which resulted in use of the PIC
variant of the 64-bit atomics even when this was not strictly
necessary.  Since this implementation is more constrained with respect
to register use, it is likely to produce less efficient code on
average.  So, the intent of this commit is to "tighten up" the logic
which decides when to activate the PIC variant.

This commit changes "case (2)" from:
  (__PIC__ is defined) &&
     (compiler is known to reserve EBX)
to
  (__PIC__ is defined) &&
    (__PIC__ was NOT defined at configure time) &&
    (the compiler is "known" to possibly reserve EBX)

In other words, if __PIC__ was defined at configure time, we trust the
logic for "case (1)" which is a configure test for availability of
EBX.  So, with a new gcc (which can schedule EBX) on Darwin (which is
PIC by default) we now use the "normal" 64-bit atomics instead of the
variant which manually saves and restores EBX.

NOTE1: Prior to the gcc-5 changes, our heuristic had actually been
producing "exact" decisions and therefore we had not considered the
fact that the PIC variant of the 64-bit atomics would actually be
incorrect if used with a compiler that was capable of scheduling EBX
as a normal register.  So, another recent commit has revised the
corresponding inline asm to be safe/correct regardless of whether or
not EBX can be schduled by the compiler.  It is now safe to be
"conservative" (e.g. favoring the PIC variant) in our heuristic.  Thus
this present commit is NOT required for correctness.

NOTE2: If we are really concerned with avoiding the PIC variant of the
64-bit atomics on x86, then in the future we could replace this block
of logic with a configure time test that checks whether or not EBX is
reserved when one passes the compiler flags to force PIC.  While that
should provide "perfect" decisions, it will require that the configure
test know the proper flags, which depend on the compiler and on the
OS.